requirements.txt
Django==1.4.3
evernote==1.25.0
httplib2==0.7.6
oauth2==1.5.211
wsgiref==0.1.2
django-bootstrap-form==2.0.5
pillow
